Tests/LibWeb/Text/expected/UIEvents/pointer-events.txt

pointerdown at (10, 20)
mousedown at (10, 20)
pointerup at (10, 20)
mouseup at (10, 20)
